fix(bedrock): report one set of token counters across both doors
Four counter defects shared a cause: each path built GenerationInfo by
hand. The legacy stream wrote the uncached input as PromptTokens and never
wrote the cache keys at all, because the nested usage struct it decoded
had only two fields; it also read the terminal counters from message.usage,
which the message_delta event does not carry. Both Bedrock doors reported
every counter as int32 while anthropic, openai and googleai report int, so
examples/reasoning-tokens read the type assertion as "no thinking tokens"
on Bedrock alone.

Each door now folds its usage through one mapper, and the streamed message
type is named so a field added to it cannot break the tests that build it.

PAGI-157

+++ llms/bedrock/converse_stream_usage_test.go
@@ -45,10 +45,10 @@ func TestConverseStreamReportsTheSameCountersAsTheWholeAnswer(t *testing.T) {
 	require.Len(t, resp.Choices, 1)
 
 	info := resp.Choices[0].GenerationInfo
-	assert.Equal(t, int32(304), info["CompletionTokens"])
-	assert.Equal(t, int32(3919), info["TotalTokens"])
-	assert.Equal(t, int32(3602), info["CacheReadInputTokens"])
-	assert.Equal(t, int32(3602), info["PromptCachedTokens"])
-	assert.Equal(t, int32(300), info["CacheCreationInputTokens"])
-	assert.Equal(t, int32(3915), info["PromptTokens"])
+	assert.Equal(t, 304, info["CompletionTokens"])
+	assert.Equal(t, 3919, info["TotalTokens"])
+	assert.Equal(t, 3602, info["CacheReadInputTokens"])
+	assert.Equal(t, 3602, info["PromptCachedTokens"])
+	assert.Equal(t, 300, info["CacheCreationInputTokens"])
+	assert.Equal(t, 3915, info["PromptTokens"])
 }
